+++ addOns/postman/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Allow Automation Framework users to limit the number of Postman messages to import (`maxMessages`). Ex: If testing authentication, access, etc.
+
 ### Changed
 - Maintenance changes.